fix(tf-docs): run docs pre-commit prior to init/validate to set versions correctly (#14)
# What

Updated the pre-commit config to run the terraform-docs step prior to running the other terraform steps

# Why

If init & validate are run prior to running terraform-docs the tool will pull in the precise provider versions, rather than what is specified in the versions.tf. Reordering the execution of these pre-commit steps forces terraform-docs to pull the provider versions from the versions.tf, which is more useful for the documentation and causes less issues in diffs every time the provider updates.